Put the debug related commands into the debug lib instead of just floating in server.cpp
